Add CollectionManager
Add a collection manager so that code responsible for setting up MongoDB collections is centralized.
