+++ src/Controllers/UsersController.cs
@@ -12,13 +12,11 @@ namespace ChatApi.Controllers
     [Route("api/[controller]")]
     public class UsersController : Controller
     {
-        private string connString = Environment.GetEnvironmentVariable("MONGODB_CONN_STR");
-
         // GET api/users
         [HttpGet]
         public IActionResult Get()
         {
-            var collection = GetUserCollection();
+            var collection = CollectionManager.GetUserCollection();
             var users = collection.Find(u => true).ToList();
             return Json(users);
         }
@@ -27,7 +25,7 @@ namespace ChatApi.Controllers
         [HttpGet("{id}")]
         public IActionResult Get(string id)
         {
-            var collection = GetUserCollection();
+            var collection = CollectionManager.GetUserCollection();
             var user = collection.Find(u => u.Id == id).FirstOrDefault();
 
             if (user == null)
@@ -43,7 +41,7 @@ namespace ChatApi.Controllers
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
 
-            var collection = GetUserCollection();
+            var collection = CollectionManager.GetUserCollection();
             user.Id = ObjectId.GenerateNewId().ToString();
             collection.InsertOne(user);
             return Get(user.Id);
@@ -58,7 +56,7 @@ namespace ChatApi.Controllers
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
 
-            var collection = GetUserCollection();
+            var collection = CollectionManager.GetUserCollection();
             collection.ReplaceOne(u => u.Id == id, user);
             return Get(user.Id);
         }
@@ -67,17 +65,9 @@ namespace ChatApi.Controllers
         [HttpDelete("{id}")]
         public IActionResult Delete(string id)
         {
-            var collection = GetUserCollection();
+            var collection = CollectionManager.GetUserCollection();
             collection.DeleteOne(u => u.Id == id);
             return Ok();
-        }
-
-        private IMongoCollection<User> GetUserCollection()
-        {
-            var client = new MongoClient(connString);
-            var db = client.GetDatabase("chat-demo");
-            var collection = db.GetCollection<User>("users");
-            return collection;
         }
     }
 }
